+++ dartsim/src/SimulationFeatures.cc
@@ -21,6 +21,7 @@
 #include "SimulationFeatures.hh"
 
 #include "ignition/common/Profiler.hh"
+#include "ignition/physics/GetContacts.hh"
 
 namespace ignition {
 namespace physics {
@@ -82,8 +83,14 @@ SimulationFeatures::GetContactsFromLastStep(const Identity &_worldID) const
       std::size_t shape2ID =
           this->shapes.IdentityOf(dtShapeFrame2->asShapeNode());
 
-      // TODO(addisu) Add normal, depth and wrench to extraData.
       CompositeData extraData;
+
+      // Add normal, depth and wrench to extraData.
+      auto &extraContactData = extraData.Get<ExtraContactData>();
+      extraContactData.force = dtContact.force;
+      extraContactData.normal = dtContact.normal;
+      extraContactData.depth = dtContact.penetrationDepth;
+
       outContacts.push_back(
           {this->GenerateIdentity(shape1ID, this->shapes.at(shape1ID)),
            this->GenerateIdentity(shape2ID, this->shapes.at(shape2ID)),